Fixed small issues with loading data from TowerData Adjusted Enemy stats (prefabs) to take more than 1 hit from towers Adjusted Enemy spawning XML file so more enemies spawn - bring on the pain

+++ Assets/Code/Towers/Tower Data/TowerData.cs
@@ -24,7 +24,7 @@ public class TowerData
     public float ThraceiumDamage;
     public float BallisticDamage;
     public float TransitionTime;
-    public float InstallCost;
+    public int InstallCost;
     public float MaintenanceCost;
 	public float StartupTime;
 	public float TrackingSpeed;
